backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.4.0
spacy==3.7.2

# Resume Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Database (optional)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
